requirements.txt
flask
requests
Flask-Cors
google-cloud-firestore